+++ history.md
@@ -1,3 +1,8 @@
+## Edge
+
+* Add more helpers to `wti init` a project
+* Add ability to add a source locale to a project from the command line: the first `wti addlocale xx` sets the source locale.
+
 ## Version 2.0.0.rc4 / 2012-04-23
 
 * Allow instantiating `WebTranslateIt::String` and `WebTranslateIt::Term` using a Hash made of symbols or strings.